GromovWitten

Documentation for GromovWitten.

GromovWitten.GromovWitten

GromovWitten.constterm

GromovWitten.feynman_integral_branchtype

GromovWitten.feynman_integral_branchtype_order

GromovWitten.feynman_integral_degree

GromovWitten.feynman_integral_degree_order

GromovWitten.feynman_integral_degree_sum

GromovWitten.feynman_integral_degree_sum_order

GromovWitten.filter_term

GromovWitten.flip_signature

GromovWitten.inv_sfunction

GromovWitten.loopterm

GromovWitten.partition

GromovWitten.proterm

GromovWitten.sfunction

GromovWitten.signature_and_multiplicities

GromovWitten.substitute

GromovWitten.sum_of_coeff

GromovWitten.GromovWitten — Module

GromovWitten is a package for computing Gromov-Witten invariant via Feynman Integral .

GromovWitten.constterm — Method

constterm( x1::QQMPolyRingElem, x2::QQMPolyRingElem, N::Integer) returns the constant term of the propagator

Examples (without vertex contribution)

julia> constterm(x[1],x[2],3)

3x[1]^6 + 2x[1]^5x[2] + x[1]^4x[2]^2

constterm(x1::QQMPolyRingElem, x2::QQMPolyRingElem, z1::QQMPolyRingElem, z2::QQMPolyRingElem,aa::Integer, N::Integer)

Examples (without vertex contribution)

here aa=1 is the order for sfunction series and $N=\sum_{n=1}^{3g-3} a_i$ where a=\[a_1,…,a_{3g-3}\] is a branche type.

julia> constterm(x[1],x[2],z[1],z[2],1,2)

1//18x[1]^4z[1]^2z[2]^2 + 1//3x[1]^4z[1]^2 + 1//3x[1]^4z[2]^2 + 2x[1]^4

+ 1//576x[1]^3x[2]z[1]^2z[2]^2 + 1//24x[1]^3x[2]z[1]^2 + 1//24x[1]^3x[2]z[2]^2 + x[1]^3*x[2]

GromovWitten.feynman_integral_branchtype — Method

feynmanintegralbranchtype( x::Vector{QQMPolyRingElem}, q::Vector{QQMPolyRingElem}, G::graph ,a::Vector{Int64} ;l=zeros(Int,nv(G)))

compute the Feynman Integral for a specified branch type a for all ordering Ω

Examples (without vertex contribution)

julia> G=graph(ve)

graph([(1, 3), (1, 2), (1, 2), (2, 4), (3, 4), (3, 4)])

julia> a=[0,2,1,0,0,1];

julia> feynmanintegralbranchtype(x,q,G,a)

256q[2]^2q[3]*q[6]

feynmanintegralbranchtype(x::Vector{QQMPolyRingElem}, q::Vector{QQMPolyRingElem},z::Vector{QQMPolyRingElem}, G::graph ,a::Vector{Int64} ;aa=0,l=zeros(Int,nv(G)),g=zeros(Int,nv(G)))

Examples (with vertex contribution)

julia> G=graph(ve) graph([(1, 2), (2, 3), (1, 3)])

julia> a=[0,0,3];

julia> feynmanintegralbranchtype(x,q,z,G,a,aa=1,g=[1,0,0])

115//3*q[3]^3

GromovWitten.feynman_integral_branchtype_order — Method

feynmanintegralbranchtype_order( x::Vector{QQMPolyRingElem}, q::Vector{QQMPolyRingElem}, G::graph ,a::Vector{Int64} ,Ω::Vector{Int64};l=zeros(Int,nv(G)))

compute the Feynman Integral for a specified branch type a for a fixed ordering Ω

Examples (without vertex contribution)

julia> G=graph(ve)

graph([(1, 3), (1, 2), (1, 2), (2, 4), (3, 4), (3, 4)])

julia> a=[0,2,1,0,0,1];

julia> Ω=[1,3,4,2];

julia> feynmanintegralbranchtype_order(x,q,G,a,Ω)

128q[2]^2q[3]*q[6]

feynmanintegralbranchtype_order(x::Vector{QQMPolyRingElem}, q::Vector{QQMPolyRingElem},z::Vector{QQMPolyRingElem}, G::graph ,a::Vector{Int64}, Ω::Vector{Int64};aa=0,l=zeros(Int,nv(G)),g=zeros(Int,nv(G)))

Examples (with vertex contribution)

julia> G=graph(ve) graph([(1, 2), (2, 3), (1, 3)])

julia> a=[0,0,3];

julia> Ω=[1,2,3];

julia> feynmanintegralbranchtype_order(x,q,z,G,a,Ω,aa=1,g=[1,0,0])

115//6*q[3]^3

GromovWitten.feynman_integral_degree — Method

feynmanintegraldegree(x::Vector{QQMPolyRingElem}, q::Vector{QQMPolyRingElem}, G::graph,d::Integer;l=zeros(Int,nv(G)))

compute the Feynman Integral for over all the partitions of the degree d for all ordering Ω

Examples (without vertex contribution)

julia> G=graph(ve)

graph([(1, 3), (1, 2), (1, 2), (2, 4), (3, 4), (3, 4)])

julia> a=[0,2,1,0,0,1];

julia> feynmanintegraldegree(x,q,G,3)

288q[1]^3 + 32q[1]^2q[2] + 32q[1]^2q[3] + 32q[1]^2q[5] + 32q[1]^2q[6] + 8q[1]q[2]q[5] + 8q[1]q[2]q[6] + 8q[1]q[3]q[5] + 8q[1]q[3]q[6] + 24q[2]^3 + 152q[2]^2q[3] + 8q[2]^2q[5] + 8q[2]^2q[6] + 152q[2]q[3]^2 + 32q[2]q[3]q[5] + 32q[2]q[3]q[6] + 32q[2]q[4]^2 + 8q[2]q[4]q[5] + 8q[2]q[4]q[6] + 8q[2]q[5]^2 + 32q[2]q[5]q[6] + 8q[2]q[6]^2 + 24q[3]^3 + 8q[3]^2q[5] + 8q[3]^2q[6] + 32q[3]q[4]^2 + 8q[3]q[4]q[5] + 8q[3]q[4]q[6] + 8q[3]q[5]^2 + 32q[3]q[5]q[6] + 8q[3]q[6]^2 + 288q[4]^3 + 32q[4]^2q[5] + 32q[4]^2q[6] + 24q[5]^3 + 152q[5]^2q[6] + 152q[5]q[6]^2 + 24q[6]^3

feynmanintegraldegree( x::Vector{QQMPolyRingElem}, q::Vector{QQMPolyRingElem},z::Vector{QQMPolyRingElem}, G::graph,d::Integer;aa=0,l=zeros(Int,nv(G)),g=zeros(Int,nv(G)))

Examples (with vertex contribution)

julia> G=graph(ve) graph([(1, 2), (2, 3), (1, 3)])

julia> a=[0,0,3];

julia> feynmanintegraldegree(x,q,z,G,3,aa=1,g=[1,0,0])

115//3q[1]^3 + 1//4q[1]^2q[2] + 1//4q[1]^2q[3] + 1//4q[1]q[2]^2 + 1//2q[1]q[2]q[3] + 1//4q[1]q[3]^2 + 115//3q[2]^3 + 1//4q[2]^2q[3] + 1//4q[2]q[3]^2 + 115//3q[3]^3

GromovWitten.feynman_integral_degree_order — Method

feynmanintegraldegree_order( x::Vector{QQMPolyRingElem}, q::Vector{QQMPolyRingElem}, G::graph,o::Vector{Int64},d::Integer;aa=0,l=zeros(Int,nv(G)))

compute the Feynman Integral for over all the partitions of the degree d for a fixed ordering Ω

Examples (without vertex contribution)

julia> G=graph(ve)

graph([(1, 3), (1, 2), (1, 2), (2, 4), (3, 4), (3, 4)])

julia> a=[0,2,1,0,0,1];

julia> Ω=[1,3,4,2];

julia> feynmanintegraldegree_order(x,q,G,Ω,4)

4q[1]^2q[2]q[3] + 4q[1]q[2]^2q[5] + 4q[1]q[2]^2q[6] + 4q[1]q[3]^2q[5] + 4q[1]q[3]^2q[6] + 176q[2]^4 + 496q[2]^3q[3] + 60q[2]^3q[5] + 60q[2]^3q[6] + 788q[2]^2q[3]^2 + 128q[2]^2q[3]q[5] + 128q[2]^2q[3]q[6] + 4q[2]^2q[4]q[5] + 4q[2]^2q[4]q[6] + 16q[2]^2q[5]^2 + 16q[2]^2q[6]^2 + 496q[2]q[3]^3 + 128q[2]q[3]^2q[5] + 128q[2]q[3]^2q[6] + 4q[2]q[3]q[4]^2 + 48q[2]q[3]q[5]^2 + 4q[2]q[3]q[5]q[6] + 48q[2]q[3]q[6]^2 + 176q[3]^4 + 60q[3]^3q[5] + 60q[3]^3q[6] + 4q[3]^2q[4]q[5] + 4q[3]^2q[4]q[6] + 16q[3]^2q[5]^2 + 16q[3]^2q[6]^2

feynmanintegraldegree_order( x::Vector{QQMPolyRingElem}, q::Vector{QQMPolyRingElem},z::Vector{QQMPolyRingElem}, G::graph,o::Vector{Int64},d::Integer;aa=0,l=zeros(Int,nv(G)),g=zeros(Int,nv(G)))

Examples (with vertex contribution)

julia> G=graph(ve) graph([(1, 2), (2, 3), (1, 3)])

julia> a=[0,0,3];

julia> Ω=[1,2,3];

julia> feynmanintegraldegree_order(x,q,z,G,o,3,aa=1,g=[1,0,0])

1//24q[1]^2q[2] + 1//24q[1]^2q[3] + 1//24q[1]q[2]^2 + 1//12q[1]q[2]q[3] + 1//24q[1]q[3]^2 + 1//24q[2]^2q[3] + 1//24q[2]q[3]^2 + 115//6q[3]^3

GromovWitten.feynman_integral_degree_sum — Method

feynmanintegraldegree_sum(x::Vector{QQMPolyRingElem}, q::Vector{QQMPolyRingElem}, G::graph, d::Union{Int64, Vector{Int64}}; l=zeros(Int, nv(G)))

compute the sum of all Feynman Integrals up to a certain degree d for all ordering Ω

Examples (without vertex contribution)

julia> G=graph(ve)

graph([(1, 3), (1, 2), (1, 2), (2, 4), (3, 4), (3, 4)])

julia> feynmanintegraldegree_sum(x,q,G,3)

288q[1]^3 + 32q[1]^2q[2] + 32q[1]^2q[3] + 32q[1]^2q[5] + 32q[1]^2q[6] + 8q[1]^2 + 8q[1]q[2]q[5] + 8q[1]q[2]q[6] + 8q[1]q[3]q[5] + 8q[1]q[3]q[6] + 24q[2]^3 + 152q[2]^2q[3] + 8q[2]^2q[5] + 8q[2]^2q[6] + 152q[2]q[3]^2 + 32q[2]q[3]q[5] + 32q[2]q[3]q[6] + 8q[2]q[3] + 32q[2]q[4]^2 + 8q[2]q[4]q[5] + 8q[2]q[4]q[6] + 8q[2]q[5]^2 + 32q[2]q[5]q[6] + 8q[2]q[6]^2 + 24q[3]^3 + 8q[3]^2q[5] + 8q[3]^2q[6] + 32q[3]q[4]^2 + 8q[3]q[4]q[5] + 8q[3]q[4]q[6] + 8q[3]q[5]^2 + 32q[3]q[5]q[6] + 8q[3]q[6]^2 + 288q[4]^3 + 32q[4]^2q[5] + 32q[4]^2q[6] + 8q[4]^2 + 24q[5]^3 + 152q[5]^2q[6] + 152q[5]q[6]^2 + 8q[5]q[6] + 24q[6]^3

feynmanintegraldegree_sum(x::Vector{QQMPolyRingElem}, q::Vector{QQMPolyRingElem},z::Vector{QQMPolyRingElem}, G::graph, d::Union{Int64, Vector{Int64}}; aa=0, l=zeros(Int, nv(G)), g=zeros(Int, nv(G)))

Examples (with vertex contribution)

julia> G=graph(ve) graph([(1, 2), (2, 3), (1, 3)])

julia> feynmanintegraldegree_sum(x,q,z,G,3,aa=1,g=[1,0,0])

115//3q[1]^3 + 1//4q[1]^2q[2] + 1//4q[1]^2q[3] + 19//4q[1]^2 + 1//4q[1]q[2]^2 + 1//2q[1]q[2]q[3] + 1//4q[1]q[2] + 1//4q[1]q[3]^2 + 1//4q[1]q[3] + 1//12q[1] + 115//3q[2]^3 + 1//4q[2]^2q[3] + 19//4q[2]^2 + 1//4q[2]q[3]^2 + 1//4q[2]q[3] + 1//12q[2] + 115//3q[3]^3 + 19//4q[3]^2 + 1//12q[3]

GromovWitten.feynman_integral_degree_sum_order — Method

feynmanintegraldegreesumorder(x::Vector{QQMPolyRingElem}, q::Vector{QQMPolyRingElem}, G::graph,o::Vector{Int64}, d::Union{Int64, Vector{Int64}}; aa=0, l=zeros(Int, nv(G)), g=zeros(Int, nv(G)))

compute the sum of all Feynman Integrals up to a certain degree d with a fixed ordering Ω

Examples (without vertex contribution)

julia> G=graph(ve)

graph([(1, 3), (1, 2), (1, 2), (2, 4), (3, 4), (3, 4)])

julia> a=[0,2,1,0,0,1];

julia> Ω=[1,3,4,2];

julia> feynmanintegraldegreesumorder(x,q,G,Ω,3) 12q[2]^3 + 76q[2]^2q[3] + 4q[2]^2q[5] + 4q[2]^2q[6] + 76q[2]q[3]^2 + 16q[2]q[3]q[5] + 16q[2]q[3]q[6] + 4q[2]q[3] + 12q[3]^3 + 4q[3]^2q[5] + 4q[3]^2q[6] –––––––––––––

feynmanintegraldegreesumorder(x::Vector{QQMPolyRingElem}, q::Vector{QQMPolyRingElem},z::Vector{QQMPolyRingElem}, G::graph,o::Vector{Int64}, d::Union{Int64, Vector{Int64}}; aa=0, l=zeros(Int, nv(G)), g=zeros(Int, nv(G)))

Examples (with vertex contribution)

julia> G=graph(ve) graph([(1, 2), (2, 3), (1, 3)])

julia> a=[0,0,3];

julia> Ω=[1,2,3];

julia> feynmanintegraldegreesumorder(x,q,z,G,o,3,aa=1,g=[1,0,0])

1//24q[1]^2q[2] + 1//24q[1]^2q[3] + 1//24q[1]q[2]^2 + 1//12q[1]q[2]q[3] + 1//24q[1]q[2] + 1//24q[1]q[3]^2 + 1//24q[1]q[3] + 1//24q[2]^2q[3] + 1//24q[2]q[3]^2 + 1//24q[2]q[3] + 115//6q[3]^3 + 19//8q[3]^2 + 1//24q[3]

GromovWitten.filter_term — Method

filter_term(p::Union{QQMPolyRingElem, Int64}, variables::Vector{QQMPolyRingElem}, s::Vector{Int64})

replaces all terms of the polynomial p with zero whenever the variables raised to a power of s1 exceed the specified power s.

Examples

julia> p=feynmanintegraldegree(x,q,G,4)

8q[1]^3q[2] + 8q[1]^3q[3] + 8q[1]^3q[4] + 54q[1]^2q[2]^2 + 18q[1]^2q[2]*q[3]

18q[1]^2q[2]q[4] + 54q[1]^2q[3]^2 + 18q[1]^2q[3]q[4] + 54q[1]^2q[4]^2

56q[1]q[2]^3 + 6q[1]q[2]^2q[3] + 6q[1]q[2]^2q[4] + 6q[1]q[2]*q[3]^2

12q[1]q[2]q[3]q[4] + 6q[1]q[2]q[4]^2 + 56q[1]q[3]^3 + 6q[1]q[3]^2q[4]

6q[1]q[3]q[4]^2 + 56q[1]*q[4]^3

we replace all term in $p$ with q[1]^a*q[2]^b*q[3]^c > q[1]*q[2]*q[3] by zero,this means all power $(a,b,c)>(1,1,1)$

julia> filter_term(p,[q[1],q[2],q[3]],[1,1,1])

12q[1]q[2]q[3]q[4] + 6q[1]q[2]q[4]^2 + 6q[1]q[3]q[4]^2 + 56q[1]q[4]^3

GromovWitten.flip_signature — Method

flip_signature(G::graph ,p::Vector{Int64},a::Vector{Int64})

Let Ω=[x1,...,xn] be a given Order and $a$ a branche type,flip_signature returns -1 if xi<xj and O else. It will return -2 in case the Graph G has a loop.

GromovWitten.inv_sfunction — Method

inv_sfunction(z::QQMPolyRingElem,aa::Int64) returns the inverse sfunction

\[ \frac{1}{S(z,aa)}=\frac{z}{2 Sinh(z/2)}= \sum_{n = 0}^{aa} \left( \left\{\begin{array}{ll} 1 & if && n = 1\\ - \frac{- 2^n (- 2 + 2^n)}{n!} B_n & (n > = 1 && (- 1 + n)\mod 2 = 1) \end{array}\right. \right) z^n \]

Where $B_n$ is Bernoulli number and ${aa} \rightarrow \infty$.

Examples

julia> inv_sfunction(x[1],4)

7//5760x[1]^4 - 1//24x[1]^2 + 1

GromovWitten.loopterm — Method

loopterm( z::QQMPolyRingElem, q::QQMPolyRingElem, aa::Integer, a::Integer)

returns loop contribution with nonzero genus gi at a vertex i.

GromovWitten.partition — Method

partition(k::Integer, n::Integer)

Note:This function returns the number of partitions of $n$ into fixed $k$ parts.

Examples

julia> partition(3,4)

15-element Vector{Vector{Int64}}:

[4, 0, 0] [3, 1, 0] [3, 0, 1] [2, 2, 0] [2, 1, 1] [2, 0, 2] [1, 3, 0] [1, 2, 1] [1, 1, 2] [1, 0, 3] [0, 4, 0] [0, 3, 1] [0, 2, 2] [0, 1, 3] [0, 0, 4]

GromovWitten.proterm — Method

proterm( x1::QQMPolyRingElem, x2::QQMPolyRingElem, q::QQMPolyRingElem, a::Integer, N::Integer)

returns the non constant term of the propagator

Examples (without vertex contribution)

julia> proterm(x[1],x[2],q[1],1,2)

x[1]^3x[2]q[1] + x[1]x[2]^3q[1]

Examples (with vertex contribution)

proterm( x1::QQMPolyRingElem, x2::QQMPolyRingElem,z1::QQMPolyRingElem, z2::QQMPolyRingElem, q::QQMPolyRingElem, a::Integer,aa::Integer, N::Integer) julia> proterm(x[1],x[2],z[1],z[2],q[1],1,1,2)

1//576x[1]^3x[2]q[1]z[1]^2z[2]^2 + 1//24x[1]^3x[2]q[1]*z[1]^2

1//24x[1]^3x[2]q[1]z[2]^2 + x[1]^3x[2]q[1] + 1//576x[1]x[2]^3q[1]z[1]^2*z[2]^2

1//24x[1]x[2]^3q[1]z[1]^2 + 1//24x[1]x[2]^3q[1]z[2]^2 + x[1]x[2]^3q[1]

GromovWitten.sfunction — Method

sfunction(z::QQMPolyRingElem,k::Int64)

Note:The function sfunction(z,k) takes account vertex contributions.

\[S(z, {aa}) = \sum_{n = 0}^{{aa}} \dfrac{2^{- 1 - n} (1 + (- 1)^n) }{(n + 1) !} z^n = \sum_{n = 0}^{{aa}} \dfrac{{2^{- 2 n}} }{(2 n + 1) !} z^n, {aa} \rightarrow \infty\]

Examples

julia> sfunction(x[1],4)

1//92897280z[1]^8 + 1//322560z[1]^6 + 1//1920z[1]^4 + 1//24z[1]^2

GromovWitten.signature_and_multiplicities — Method

signature_and_multiplicities( G::graph, a::Vector{Int64})

returns flip_signature and their multiplicities.

Examples

julia> G=graph(ve)

graph([(1, 1), (1, 2), (2, 3), (3, 1)])

a=[2,0,0,1]

julia> signature_and_multiplicities(G,a)

4-element Vector{Tuple{Int64, Vector{Int64}}}:

(1, [-2, 0, 0, 1]) (2, [-2, -1, 0, 1]) (2, [-2, 0, -1, 1]) (1, [-2, -1, -1, 1])

GromovWitten.substitute — Method

substitute(q::Vector{QQMPolyRingElem},p::Union{QQMPolyRingElem, Int64})

replace all the variables by the first variable of p. With x=\[x_1,x2,x3 \] and p(x_1,x2,x3), substitute(x,p) returns p(x1,x_1,x_1)

julia> f=x[1]x[2]+x[1]^3x[2]+5x[1]^6-2x[3]x[2] 5x[1]^6 + x[1]^3x[2] + x[1]x[2] - 2x[2]x[3]

julia> substitute(x,f)

5*x[1]^6 + x[1]^4 - x[1]^2

GromovWitten.sum_of_coeff — Method

sumofcoeff(p::QQMPolyRingElem)

compute the sum of coefficient of the polynomial p.

julia> f=3x[1]^6 + 2x[1]^5x[2] + x[1]^4x[2]^2

julia> sumofcoeff(f)

6

To run:
- pull the package in your local disk
In the local folder where the package is located, type in the terminal.
- julia --project
-  using GromovWitten
-  ve=[(1, 2), (1,2),(2, 4), (1, 3) ,(3, 4),(3,4)]
-   G=graph(ve)
-   R,x,q=polynomial_ring(G,"x","q")
-  a=[0,2,1,0,0,1]
-  feynman_integral_branchtype(x,q,G,a)
-  feynman_integral_degree(x,q,G ,4)
-  sum_of_coeff(feynman_integral_degree(x,q,G ,4))