GromovWitten

Documentation for tropicalfeynman.

GromovWitten.GromovWitten

GromovWitten.Sfunction

GromovWitten.partition

GromovWitten.GromovWitten — Module

GromovWitten is a package for computing Feynman Integral via Gromov-Witten invariant.

GromovWitten.Sfunction — Method

Sfunction(z::QQMPolyRingElem,k::Int64) Note:The function Sfunction(z,k) takes account vertex contributions. $S (z, {aa}) = \sum_{n = 0}^{{aa}} \dfrac{2^{- 1 - n} (1 + (- 1)^n) }{(n + 1) !} z^n = \sum_{n = 0}^{{aa}} \dfrac{{2^{- 2 n}} }{(2 n + 1) !} z^n, {aa} \rightarrow \infty$

Examples

jldoctest julia> Sfunction(z[1],4) 1//92897280*z[1]^8 + 1//322560*z[1]^6 + 1//1920*z[1]^4 + 1//24*z[1]^2`

GromovWitten.partition — Method

partition(k::Integer, n::Integer)

Note:This function returns the number of partitions of $n$ into fixed $k$ parts.

Examples

jldoctest julia> partition(3,4) 15-element Vector{Vector{Int64}}: [4, 0, 0] [3, 1, 0] [3, 0, 1] [2, 2, 0] [2, 1, 1] [2, 0, 2] [1, 3, 0] [1, 2, 1] [1, 1, 2] [1, 0, 3] [0, 4, 0] [0, 3, 1] [0, 2, 2] [0, 1, 3] [0, 0, 4]`

To run:
- pull the package in your local disk
In the local folder where the package is located, type in the terminal.
- julia --project
-  using GromovWitten
-  ve=[(1, 2), (1,2),(2, 4), (1, 3) ,(3, 4),(3,4)]
-   G=graph(ve)
-   R,x,q=polynomial_ring(G,"x","q")
-  a=[0,2,1,0,0,1]
-  feynman_integral_branchtype(x,q,G,a)
-  feynman_integral_degree(x,q,G ,4)
-  subt(R,x,q,feynman_integral_degree(x,q,G ,4))